requirements.txt
fastapi
pydantic
langgraph
uvicorn
python-dotenv
qdrant-client
openai
pillow
numpy
